+++ utils.py
@@ -1,0 +1,14 @@
+
+def build_menu(buttons, n_cols, header_buttons=None, footer_buttons=None):
+    menu = []
+    for i in range(0, len(buttons)):
+        item = buttons[i]
+        if i % n_cols == 0:
+            menu.append([item])
+        else:
+            menu[int(i / n_cols)].append(item)
+    if header_buttons:
+        menu.insert(0, header_buttons)
+    if header_buttons:
+        menu.append(footer_buttons)
+    return menu